backend/requirements.txt
fastapi==0.104.1 # For building the API
uvicorn==0.24.0 # For ASGI server
sqlalchemy==2.0.23 # For database ORM
aiosqlite==0.19.0  # For SQLite database
python-jose[cryptography]==3.3.0 # For JWT authentication
passlib[bcrypt]==1.7.4 # For password hashing
python-multipart==0.0.6 # For handling file uploads
alembic==1.11.1 # For database migrations
pymysql==1.0.3 # For MySQL database
python-dotenv==1.0.0 # For environment variable management
pytest==7.4.0  # For testing
httpx==0.24.1  # For making HTTP requests in tests
pytest-asyncio==0.21.0  # For async tests
fastapi-pagination==0.12.2  # For pagination support
databases[sqlite]==0.9.0  # For async database support with SQLite
fastapi-jwt-auth==0.5.0  # For JWT authentication (pinned to available PyPI version)
pydantic[email]==2.12.5  # For email validation (bumped to satisfy pydantic-settings)
aiofiles==23.1.0  # For handling file uploads asynchronously
python-crontab==2.6.0  # For scheduling tasks
bcrypt==4.1.2  # For password hashing
itsdangerous==2.1.2  # For generating secure tokens
pydantic-settings==2.12.0  # Required for BaseSettings (pydantic v2 migration)
aiomysql  # Async MySQL driver required for mysql+aiomysql SQLAlchemy URL

